feat: used information about planed car charging to build load profile
with this any planned charging is added to the profile. The information that charging will happen exist so we can use it to build a more precise prediction.
For simplicity the amount needed energy is evenly distributed over the charging hours. This is wrong if for example the late-charge option is set so that there is a pause between start and finish of the charging session.